+++ CHANGELOG.md
@@ -6,4 +6,20 @@
 
 ![Dynamic program](./asset/image/dp.jpeg) 
 
-+ 9月9日 Begin ans set up the github repo
+> working on :point_right:. done :white_check_mark:
+
++ ***9月9日*** Begin ans set up the github repo  
++ begin the ***dynamic progress*** part.  
++ :white_check_mark: 509 Fibonacci Number
++ ***9月10日*** Begin ans set up the github repo  
++ :white_check_mark: 70 Climbing Stairs
++ :white_check_mark: 746 min cost climb stairs
++ :white_check_mark: 62 unique path
++ :white_check_mark: 63 unique path 2
++ :white_check_mark: 343 integer partition
++ :white_check_mark: 96 unique bst
++ ***9月11日*** Begin ans set up the github repo  
++ :white_check_mark: 416 subset equal
++ :white_check_mark: 1049 last stone weight
++ :white_check_mark: 494 target sum
++ :point_right: 474 one and zero